requirements.txt
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
dash==2.14.1
plotly==5.17.0